fixed issues

+++ src/Modules/HR/Command/PayrollProccess.php
@@ -63,11 +63,13 @@ class PayrollProccess extends ContainerAwareCommand
     $result=shell_exec("gs -o \"".$tempDir.uniqid("file")."_%04d.pdf\" -sDEVICE=pdfwrite \"".$sourceFile."\"");
     $dir = new \DirectoryIterator($tempDir);
     foreach ($dir as $fileinfo) {
-      //Añadir formato corporativo
-      $result=shell_exec("pdftk ".$tempDir.$fileinfo->getFilename()." stamp /home/operador/nominas/plantilla_nominas.pdf output ".$ocrDir.basename($fileinfo->getFilename(), '.pdf')."_%04d.pdf");
-      unlink($tempDir.$fileinfo->getFilename());
-      //Firmar documentos
-      //pdftk file63d91db71abef_0046.pdf stamp /home/operador/nominas/plantilla_nominas.pdf output format_file63d91db71abef_0046.pdf
+      if (!$fileinfo->isDot()) {
+        //Añadir formato corporativo
+        $result=shell_exec("pdftk ".$tempDir.$fileinfo->getFilename()." stamp /home/operador/nominas/plantilla_nominas.pdf output ".$ocrDir.'format_'.basename($fileinfo->getFilename(), '.pdf')."_%04d.pdf");
+        unlink($tempDir.$fileinfo->getFilename());
+        //Firmar documentos
+        //pdftk file63d91db71abef_0046.pdf stamp /home/operador/nominas/plantilla_nominas.pdf output format_file63d91db71abef_0046.pdf
+      }
     }
     //Borrar archivo original
     unlink($sourceFile);
